# docs/intel接口文档-中文版本-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Host Service",
        "description": "主机管理和WebSocket实时通信服务",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/host/hosts/available": {
            "post": {
                "tags": [
                    "浏览器插件-主机管理"
                ],
                "summary": "查询可用主机列表",
                "description": "查询可用的主机列表，支持游标分页",
                "operationId": "query_available_hosts_api_v1_host_hosts_available_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueryAvailableHostsRequest",
                                "description": "查询可用主机列表请求参数"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_AvailableHostsListResponse_"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 400,
                                    "message": "请求参数无效",
                                    "error_code": "INVALID_PARAMS"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "HTTP 方法不允许",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 405,
                                    "message": "此接口仅支持 POST 方法，请使用 POST 请求",
                                    "error_code": "METHOD_NOT_ALLOWED"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "外部服务不可用",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 503,
                                    "message": "硬件接口调用失败，请稍后重试",
                                    "error_code": "HARDWARE_API_ERROR"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/host/hosts/retry-vnc": {
            "post": {
                "tags": [
                    "浏览器插件-主机管理"
                ],
                "summary": "获取重试 VNC 列表",
                "description": "查询需要重试的 VNC 连接列表（case_state != 2 的主机）",
                "operationId": "get_retry_vnc_list_api_v1_host_hosts_retry_vnc_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_get_retry_vnc_list_api_v1_host_hosts_retry_vnc_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_RetryVNCListResponse_"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 400,
                                    "message": "请求参数无效",
                                    "error_code": "INVALID_PARAMS"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/host/hosts/release": {
            "post": {
                "tags": [
                    "浏览器插件-主机管理"
                ],
                "summary": "释放主机",
                "description": "逻辑删除指定用户的主机执行日志记录并更新主机状态（设置 del_flag = 1，host_state = 0）",
                "operationId": "release_hosts_api_v1_host_hosts_release_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReleaseHostsRequest",
                                "description": "释放主机请求数据"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "释放成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_ReleaseHostsResponse_"
                                },
                                "example": {
                                    "code": 200,
                                    "message": "主机释放成功",
                                    "data": {
                                        "updated_count": 3,
                                        "user_id": "user123",
                                        "host_list": [
                                            "host1",
                                            "host2",
                                            "host3"
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 400,
                                    "message": "主机ID格式无效",
                                    "error_code": "INVALID_HOST_ID"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/host/hosts/reset": {
            "post": {
                "tags": [
                    "浏览器插件-主机管理"
                ],
                "summary": "测试重置主机",
                "description": "重置主机状态为有效状态并删除执行日志（测试用）",
                "operationId": "reset_host_for_test_api_v1_host_hosts_reset_post",
                "parameters": [
                    {
                        "name": "accept-language",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Accept-Language 请求头",
                            "title": "Accept-Language"
                        },
                        "description": "Accept-Language 请求头"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetHostForTestRequest",
                                "description": "测试重置主机请求参数"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "重置成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Result_ResetHostForTestResponse_"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 400,
                                    "message": "主机ID格式无效",
                                    "error_code": "INVALID_HOST_ID"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "主机不存在",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 53001,
                                    "message": "主机不存在: 123",
                                    "error_code": "HOST_NOT_FOUND"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "QueryAvailableHostsRequest": {
                "properties": {
                    "tc_id": {
                        "type": "string",
                        "title": "Tc Id",
                        "description": "测试用例ID"
                    },
                    "cycle_name": {
                        "type": "string",
                        "title": "Cycle Name",
                        "description": "测试周期名称"
                    },
                    "user_name": {
                        "type": "string",
                        "title": "User Name",
                        "description": "用户名"
                    },
                    "page_size": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 1,
                        "title": "Page Size",
                        "description": "每页数量（1-100）",
                        "default": 20
                    },
                    "last_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Id",
                        "description": "上一页最后一条记录的 id。首次请求为 null，后续请求需要传入上一页最后一条记录的 host_rec_id"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email",
                        "description": "用户邮箱（可选）。如果提供，将直接使用该 email 进行外部接口认证，不查询数据库"
                    }
                },
                "type": "object",
                "required": [
                    "tc_id",
                    "cycle_name",
                    "user_name"
                ],
                "title": "QueryAvailableHostsRequest",
                "description": "查询可用主机列表请求模式 - 使用游标分页\n\n业务说明：\n- 首次请求不提供 last_id，从头开始\n- 后续请求提供上一页最后一条记录的 id（last_id）\n- 系统根据 last_id 计算内部偏移量\n- 避免多用户并发时的状态污染问题\n- 如果提供了 email，将直接使用该 email 进行外部接口认证，不查询数据库"
            },
...
            "Result_AvailableHostsListResponse_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "description": "响应码",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "响应消息",
                        "default": "操作成功"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AvailableHostsListResponse",
                        "description": "响应数据"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "响应时间戳（ISO 8601 格式）"
                    },
                    "locale": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Locale",
                        "description": "语言代码（用于多语言支持）"
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "Result[AvailableHostsListResponse]"
            },
...
        }
    }
}